LangGraph Project PipeLine (LLMOps)/pyproject.toml
[project]
name = "llmops"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.0",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-google-community>=2.0.7",
    "langchain-google-genai>=2.1.6",
    "langchain-groq>=0.3.5",
    "langchain-tavily>=0.2.7",
    "langgraph>=0.5.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "streamlit>=1.46.1",
    "uvicorn>=0.35.0",
]
